+++ CHANGELOG.md
@@ -244,7 +244,7 @@ Estos son los cambios hechos en la ECORED-BASE-2025
 - Nuevo Componente ImagenInfográfica
 - Nuevo Componente LineaTiempoD
 - Componente .tarjeta-Avatar-b tiene una variable que controla el ancho de su avatar
-- Los botones ahora tienen el nuevo esquema de colorización (.color-primario, .color-secundario, .color-acento-contenido, .color-acento-botones)
+- Los botones ahora tienen el nuevo esquema de colorización (.color-primario, .color-secundario, color-terciario, .color-terciario-claro, .color-terciario-T10.color-acento-contenido, .color-acento-botones)
 - esquema de colorización tiene color de texto basado en luminancia del color de fondo
 - Nueva distribucion de creditos
 - Nuevo componente de indicador de interacción
@@ -284,6 +284,9 @@ Estos son los cambios hechos en la ECORED-BASE-2025
   ```
   .color-primario
   .color-secundario
+  .color-terciario
+  .color-terciario-claro
+  .color-terciario-T10
   .color-acento-contenido
   .color-acento-botones
   ```